+++ instrumentation/opentelemetry-instrumentation-django/tests/test_middleware.py
@@ -121,7 +121,6 @@ class TestMiddleware(TestBase, WsgiTestBase):
         )
         self.assertEqual(span.attributes["http.scheme"], "http")
         self.assertEqual(span.attributes["http.status_code"], 200)
-        self.assertEqual(span.attributes["http.status_text"], "OK")
 
     def test_traced_get(self):
         Client().get("/traced/")
@@ -143,7 +142,6 @@ class TestMiddleware(TestBase, WsgiTestBase):
         self.assertEqual(span.attributes["http.route"], "^traced/")
         self.assertEqual(span.attributes["http.scheme"], "http")
         self.assertEqual(span.attributes["http.status_code"], 200)
-        self.assertEqual(span.attributes["http.status_text"], "OK")
 
     def test_not_recording(self):
         mock_tracer = Mock()
@@ -178,7 +176,6 @@ class TestMiddleware(TestBase, WsgiTestBase):
         self.assertEqual(span.attributes["http.route"], "^traced/")
         self.assertEqual(span.attributes["http.scheme"], "http")
         self.assertEqual(span.attributes["http.status_code"], 200)
-        self.assertEqual(span.attributes["http.status_text"], "OK")
 
     def test_error(self):
         with self.assertRaises(ValueError):