+++ instrumentation/opentelemetry-instrumentation-pyramid/tests/test_programmatic.py
@@ -35,7 +35,6 @@ def expected_attributes(override_attributes):
         "http.host": "localhost",
         "http.target": "/",
         "http.flavor": "1.1",
-        "http.status_text": "OK",
         "http.status_code": 200,
     }
     for key, val in override_attributes.items():
@@ -118,7 +117,6 @@ class TestProgrammatic(InstrumentationTest, TestBase, WsgiTestBase):
             {
                 "http.method": "POST",
                 "http.target": "/bye",
-                "http.status_text": "Not Found",
                 "http.status_code": 404,
             }
         )
@@ -137,7 +135,6 @@ class TestProgrammatic(InstrumentationTest, TestBase, WsgiTestBase):
             {
                 "http.target": "/hello/500",
                 "http.route": "/hello/{helloid}",
-                "http.status_text": "Internal Server Error",
                 "http.status_code": 500,
             }
         )