+++ instrumentation/opentelemetry-instrumentation-urllib/tests/test_urllib_integration.py
@@ -107,7 +107,6 @@ class RequestsIntegrationTestBase(abc.ABC):
                 "http.method": "GET",
                 "http.url": self.URL,
                 "http.status_code": 200,
-                "http.status_text": "OK",
             },
         )
 
@@ -158,7 +157,6 @@ class RequestsIntegrationTestBase(abc.ABC):
         span = self.assert_span()
 
         self.assertEqual(span.attributes.get("http.status_code"), 404)
-        self.assertEqual(span.attributes.get("http.status_text"), "Not Found")
 
         self.assertIs(
             span.status.status_code, trace.StatusCode.ERROR,
@@ -269,7 +267,6 @@ class RequestsIntegrationTestBase(abc.ABC):
                 "http.method": "GET",
                 "http.url": self.URL,
                 "http.status_code": 200,
-                "http.status_text": "OK",
                 "http.response.body": "Hello!",
             },
         )
@@ -299,7 +296,6 @@ class RequestsIntegrationTestBase(abc.ABC):
                 "http.method": "GET",
                 "http.url": "http://httpbin.org/status/500",
                 "http.status_code": 500,
-                "http.status_text": "Internal Server Error",
             },
         )
         self.assertEqual(span.status.status_code, StatusCode.ERROR)